fixed Loader on MainPage
